Starting the Video Player

+++ controllers/videoController.js
@@ -50,31 +50,6 @@ export const postUpload = async (req, res) => {
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
-export const videoDetail = async (req, res) => {
-  const {
-    params: { id },
-  } = req;
-  try {
-    const video = await Video.findById(id).populate("creator");
-    res.render("videoDetail", { pageTitle: video.title, video });
-    console.log(video);
-  } catch (error) {
-    res.redirect(routes.home);
-  }
-};
-
-export const getEditVideo = async (req, res) => {
-  const {
-    params: { id },
-  } = req;
-  try {
-    const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
-  } catch (error) {
-    res.redirect(routes.home);
-  }
-};
-
 export const postEditVideo = async (req, res) => {
   const {
     params: { id },
@@ -92,7 +67,42 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
-  } catch (error) {}
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
+  } catch (error) {
+    console.log(error);
+  }
   res.redirect(routes.home);
+};
+
+export const videoDetail = async (req, res) => {
+  const {
+    params: { id },
+  } = req;
+  try {
+    const video = await Video.findById(id).populate("creator");
+    res.render("videoDetail", { pageTitle: video.title, video });
+  } catch (error) {
+    res.redirect(routes.home);
+  }
+};
+
+export const getEditVideo = async (req, res) => {
+  const {
+    params: { id },
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
+  } catch (error) {
+    res.redirect(routes.home);
+  }
 };
